Restore EXE support.
Executable support was disabled because of issues with preloading/prefetching -
we really need the PDB path. What we can easily do is to try opening a session
for the executable (this is fast), and then get the PDB file path from global
symbol - and use PdbSymbolSource that does prefetch/preload optimizations.