symbols: Restore private symbols in PDB output
PublicSymbol tag does not contain symbols with private linkage, so we still need to use Data/Function tags.
However, PublicSymbol tag is important since it contains some symbols that are not available elsewhere, i.e. string constants.
Performance-wise it seems that after filtering symbols by VA reading symbols is not that much slower; it is considerably slower if we don't do Seq.distinctBy which may be the reason the original code was slow.